refactor: Improve readability metrics implementation
- Update Gunning Fog to return NaN for empty text instead of 0
- Add reliability indicator to Gunning Fog metadata
- Change grade_level type from int to float in GunningFogResult
- Follow Gunning (1952) blanket exclusion rule for hyphenated words
- Improve Coleman-Liau token filtering for alphabetic characters
- Update default tokenizer to preserve URLs and emails
- Consolidate test_gunning_fog_nlp.py into test_gunning_fog.py
- Remove internal plan file references from README

+++ pystylometry/_types.py
@@ -100,7 +100,7 @@ class GunningFogResult:
     """Result from Gunning Fog Index computation."""
 
     fog_index: float
-    grade_level: int
+    grade_level: float
     metadata: dict[str, Any]
